Rewrite spacing around colon rule with individual test cases.

+++ src/main/kotlin/format/rule/SpacingAroundColonRule.kt
@@ -1,38 +1,45 @@
 package format.rule
 
 import format.Rule
+import format.internal.isChildOfType
 import org.jetbrains.kotlin.com.intellij.lang.ASTNode
 import org.jetbrains.kotlin.com.intellij.psi.PsiWhiteSpace
 import org.jetbrains.kotlin.com.intellij.psi.impl.source.tree.LeafPsiElement
 import org.jetbrains.kotlin.com.intellij.psi.impl.source.tree.PsiWhiteSpaceImpl
-import org.jetbrains.kotlin.lexer.KtTokens.IDENTIFIER
-import org.jetbrains.kotlin.psi.KtPrimaryConstructor
+import org.jetbrains.kotlin.psi.KtClassOrObject
+import org.jetbrains.kotlin.psi.KtFunction
+import org.jetbrains.kotlin.psi.KtProperty
+import org.jetbrains.kotlin.psi.KtStringTemplateEntry
 import org.jetbrains.kotlin.psi.psiUtil.nextLeaf
 import org.jetbrains.kotlin.psi.psiUtil.prevLeaf
-import org.jetbrains.kotlin.psi.stubs.elements.KtStubElementTypes.OBJECT_DECLARATION
-import org.jetbrains.kotlin.psi.stubs.elements.KtStubElementTypes.TYPE_PARAMETER_LIST
-import org.jetbrains.kotlin.psi.stubs.elements.KtStubElementTypes.VALUE_PARAMETER
-import org.jetbrains.kotlin.psi.stubs.elements.KtStubElementTypes.VALUE_PARAMETER_LIST
 
 class SpacingAroundColonRule : Rule {
 
     override fun visit(node: ASTNode): ASTNode {
-        if (node is LeafPsiElement && node.textMatches(":")) {
+        if (node is LeafPsiElement
+                && node.textMatches(":")
+                && !node.isChildOfType(KtStringTemplateEntry::class)) {
+
             val prevLeaf = node.prevLeaf() as LeafPsiElement
             val nextLeaf = node.nextLeaf() as LeafPsiElement
 
-            val prevPrevParent = (prevLeaf.prevLeaf() as LeafPsiElement).treeParent
+            val isChildOfProperty = node.isChildOfType(KtProperty::class)
+            val isChildOfClassOrObject = node.isChildOfType(KtClassOrObject::class)
+            val isChildOfFunction = node.isChildOfType(KtFunction::class)
 
             val extraSpacingBefore = prevLeaf is PsiWhiteSpace
-                    && node.prevSibling.prevSibling !is KtPrimaryConstructor
-                    && (prevPrevParent.elementType == VALUE_PARAMETER_LIST || prevPrevParent.elementType == VALUE_PARAMETER)
+                    && (isChildOfProperty
+                    || isChildOfFunction)
 
-            val missingSpacingBefore = prevLeaf.treeParent.elementType == TYPE_PARAMETER_LIST
-                    || node.prevSibling is KtPrimaryConstructor
-                    || (prevLeaf !is PsiWhiteSpace && prevLeaf.treeParent.elementType == OBJECT_DECLARATION)
-                    || (prevLeaf.elementType == IDENTIFIER && node.treeParent.elementType != VALUE_PARAMETER)
+            val missingSpacingBefore = prevLeaf !is PsiWhiteSpace
+                    && !isChildOfProperty
+                    && !isChildOfFunction
+                    && isChildOfClassOrObject
 
-            val missingSpacingAfter = nextLeaf.elementType == IDENTIFIER
+            val missingSpacingAfter = nextLeaf !is PsiWhiteSpace
+                    && (isChildOfProperty
+                    || isChildOfClassOrObject
+                    || isChildOfFunction)
 
             if (extraSpacingBefore) {
                 prevLeaf.delete()
